#7 migrating differentiateRgb to TS
